+++ resources/assets/js/profile-directory.js
@@ -1,0 +1,4 @@
+Vue.component(
+    'profile-directory',
+    require('./components/ProfileDirectory.vue').default
+);